requirements-test.txt
pytest>=8.0
pypdf>=4.0.0
requests>=2.31
pyyaml>=6.0
